@@ -1,102 +1,55 @@
 {
   "name": "@next/extension",
   "private": true,
-  "version": "1.2.0",
-  "description": "Adiciona funcionalidades novas ao sistema de matricula da UFABC",
+  "version": "2.0.0",
+  "description": "Adiciona funcionalidades novas em sistemas da UFABC",
   "engines": {
-    "node": "^20"
+    "node": ">22"
   },
+  "type": "module",
   "scripts": {
-    "serve": "node server/app.js",
-    "start": "pnpm build && pnpm serve",
-    "deploy": "caprover deploy",
-    "build": "pnpm build:chrome",
-    "build:chrome": "gulp pack --production --vendor=chrome",
-    "build:firefox": "gulp pack --production --vendor=firefox",
-    "build:opera": "gulp pack --production --vendor=opera",
-    "build:edge": "gulp pack --production --vendor=edge",
-    "dev": "pnpm dev:chrome",
-    "staging": "gulp --watch --staging --vendor=chrome",
-    "dev:chrome": "gulp --watch --vendor=chrome",
-    "dev:firefox": "gulp --watch --vendor=firefox",
-    "dev:opera": "gulp --watch --vendor=opera",
-    "dev:edge": "gulp --watch --vendor=edge",
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "zip:firefox": "wxt zip -b firefox",
+    "compile": "vue-tsc --noEmit",
+    "postinstall": "wxt prepare",
     "fmt": "biome format ./src --write",
     "lint": "biome lint ./src"
   },
   "devDependencies": {
-    "@biomejs/biome": "^1.8.1",
-    "babel-cli": "6.x.x",
-    "babel-core": "6.x.x",
-    "babel-loader": "7.x.x",
-    "babel-preset-env": "1.7.0",
-    "chai": "4.x.x",
-    "chromereload": "0.x.x",
-    "debounce": "1.x.x",
-    "del": "3.x.x",
-    "gulp": "4.x.x",
-    "gulp-bump": "2.x.x",
-    "gulp-cache": "0.x.x",
-    "gulp-clean-css": "^3.x.x",
-    "gulp-filter": "^5.x.x",
-    "gulp-git": "^2.x.x",
-    "gulp-if": "2.x.x",
-    "gulp-imagemin": "4.x.x",
-    "gulp-json-transform": "0.x.x",
-    "gulp-less": "3.x.x",
-    "gulp-livereload": "3.x.x",
-    "gulp-plumber": "1.x.x",
-    "gulp-sass": "5.1.0",
-    "gulp-sequence": "1.x.x",
-    "gulp-sourcemaps": "^2.x.x",
-    "gulp-tag-version": "1.x.x",
-    "gulp-util": "3.x.x",
-    "gulp-zip": "^4.x.x",
-    "require-dir": "0.x.x",
-    "sass": "^1.59.3",
-    "vinyl-named": "1.x.x",
-    "webpack": "3.x.x",
-    "webpack-stream": "3.x.x",
-    "yargs": "^8.x.x"
+    "@biomejs/biome": "^1.9.3",
+    "@types/lodash-es": "^4.17.12",
+    "@types/mustache": "^4.2.5",
+    "@types/toastify-js": "^1.12.3",
+    "@wxt-dev/module-vue": "^1.0.0",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.13",
+    "typescript": "^5.5.4",
+    "unplugin-auto-import": "^0.18.3",
+    "unplugin-vue-components": "^0.27.4",
+    "vue-router": "^4.5.0",
+    "vue-tsc": "^2.0.29",
+    "wxt": "^0.19.19"
   },
   "dependencies": {
-    "autoprefixer": "^9.3.1",
-    "axios": "^0.17.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "caprover": "^1.2.0",
-    "compression": "^1.7.3",
-    "cors": "^2.8.5",
-    "css-loader": "^1.0.1",
-    "element-ui": "^2.4.11",
-    "express": "^4.16.4",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "gulp-cssimport": "^6.0.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-vueify": "^0.0.3",
-    "highcharts": "^6.2.0",
-    "is_js": "^0.9.0",
-    "jquery": "^3.2.1",
-    "lodash": "^4.17.4",
-    "material-design-icons-iconfont": "^3.0.3",
-    "mustache": "^3.0.0",
-    "postcss-import": "12.0.1",
-    "postcss-load-config": "1.0.0",
-    "postcss-loader": "1.0.0",
-    "postcss-url": "8.0.0",
-    "style-loader": "^0.23.1",
-    "toastify-js": "^1.9.0",
-    "toastr": "^2.1.4",
-    "url-loader": "^1.1.2",
-    "vue": "2.7.16",
-    "vue-loader": "^15.10.0",
-    "vue-router": "3.6.5",
-    "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "2.7.16",
-    "vue2-highcharts": "^1.2.5",
-    "vuedraggable": "^2.16.0",
-    "vueify": "^9.4.1",
-    "vueify-insert-css": "^1.0.0",
-    "vuetify": "1.2.6",
-    "webui-popover": "^1.2.18"
+    "@vueuse/core": "^11.1.0",
+    "@vueuse/integrations": "^11.1.0",
+    "change-case": "^5.4.4",
+    "dayjs": "^1.11.13",
+    "element-plus": "^2.8.4",
+    "highcharts": "^12.0.1",
+    "highcharts-vue": "^2.0.1",
+    "lodash-es": "^4.17.21",
+    "lucide-vue-next": "^0.447.0",
+    "mustache": "^4.2.0",
+    "ofetch": "^1.4.0",
+    "toastify-js": "^1.12.0",
+    "vue": "^3.5.12",
+    "vue-sonner": "^1.2.5",
+    "vuedraggable": "^4.1.0"
   }
 }
