@@ -1,1 +1,6 @@
-module.exports = {};
+export default {
+	plugins: {
+		tailwindcss: {},
+		autoprefixer: {},
+	},
+};
